+++ photomap/pm/controller/place.py
@@ -10,6 +10,7 @@ from message import success, error
 from pm.form.place import InsertPlaceForm, UpdatePlaceForm
 from pm.model.place import Place
 from pm.model.photo import Photo
+from pm.controller.authentication import is_authorized
 import logging
 import os
 from django.contrib.auth.decorators import login_required
@@ -21,6 +22,8 @@ def insert(request):
     if request.method == "POST":
         form = InsertPlaceForm(request.POST)
         if form.is_valid():
+            if not form.cleaned_data["album"].user == request.user:
+                return error("not your album")
             place = form.save()
             return success(id = place.pk)
         else:
@@ -40,6 +43,8 @@ def update(request):
             except Place.DoesNotExist: 
                 logger.warn("place %d does not exist", form.cleaned_data["id"])
                 return error("place does not exist")
+            if not is_authorized(place, request.user):
+                return error("not your place")
             form = UpdatePlaceForm(request.POST, instance = place)
             form.save()
             return success()
@@ -54,6 +59,8 @@ def delete(request):
     if request.method == "POST":
         try:
             place = Place.objects.get(pk = request.POST["id"])
+            if not is_authorized(place, request.user):
+                return error("not your place")
             place.delete()    
 
             return success()
